Fix breaking changes from dependencies
Two issues that needed to be resolved: circleci convenience images for
leiningen have removed java (lein needs java to run), and thus needed to
switch from the lein image to the openjdk-lein image.

Additionally, I had been using the long-deprecated athena-jdbc 1.x
driver, which was only available on an atlassian 3rd party repository.
The most recent checkin to that repository broke the jar integrity
(verified; the jar is not a valid zip file). No other repository had the
same or similar jar. Thus, I needed to upgrade to the 2.x jdbc driver,
which was less painful than I originally thought it would be (however,
on deployment, still need to modify the jdbc_athena_uri from http-style
get parameters (? and &), which are replaced by a literal ';'